+++ src/objects.h
@@ -1,5 +1,5 @@
 /*
- * $Id: objects.h,v 1.12 2010/01/31 02:12:25 we7u Exp $
+ * $Id: objects.h,v 1.13 2010/07/11 23:39:42 we7u Exp $
  *
  * XASTIR, Amateur Station Tracking and Information Reporting
  * Copyright (C) 1999,2000  Frank Giannandrea
@@ -26,6 +26,7 @@
 #define XASTIR_OBJECTS_H
 
 #include <X11/Intrinsic.h>
+#include <stdint.h>
 
 
 // For mutex debugging with Linux threads only
@@ -116,7 +117,7 @@ typedef struct {
     char data[PREDEFINED_OBJECT_DATA_LENGTH]; 
                                 // Data following the symbol.
     char menu_call[26];         // Name to display on menu.
-    int index;                  // Index of this object 
+    intptr_t index;             // Index of this object 
                                 // in the predefinedObjects array.
     int show_on_menu;           // !=1 to hide on menu.
     int index_of_child;         // > -1 to create two objects
